chore: update all dependencies to latest versions
Rust workspace:
- sqlx 0.7 -> 0.8 (native-tls -> rustls, fixes libssl requirement)
- rand 0.8 -> 0.10, nalgebra 0.32 -> 0.34, thiserror 1 -> 2
- Edition 2021 -> 2024 for price-solver and api-server
- Exclude market-simulation from workspace

Market simulation:
- eframe 0.31 -> 0.33, egui_plot 0.31 -> 0.34, rand 0.9 -> 0.10
- Adapt Line::new/BarChart::new to new (name, data) API
- rand::Rng -> rand::RngExt for rand 0.10

Java/Gradle:
- Gradle wrapper 9.3.1 + foojay toolchain resolver
- PMD 6.55 -> 7.14 (fixes Java 21 record parsing)
- Suppress -Xlint:classfile warnings from dependency annotations
- Fix FutureReturnValueIgnored and LongDoubleConversion warnings

Web frontend:
- Next.js 14 -> 15, React 18 -> 19, ESLint 8 -> 9
- Fix useRef typing for React 19 strict mode
- Update all npm packages to latest versions

+++ src/main/java/com/noahblclarkson/autotune/economy/LoanManager.java
@@ -389,8 +389,8 @@ public class LoanManager {
             return loan.currentBalance().divide(BigDecimal.valueOf(periodsRemaining), 2, RoundingMode.HALF_UP);
         }
 
-        double payment = balance * rate * Math.pow(1 + rate, periodsRemaining)
-                / (Math.pow(1 + rate, periodsRemaining) - 1);
+        double payment = balance * rate * Math.pow(1 + rate, (double) periodsRemaining)
+                / (Math.pow(1 + rate, (double) periodsRemaining) - 1);
 
         return BigDecimal.valueOf(payment).setScale(2, RoundingMode.HALF_UP);
     }
